moving the functions in main.py and manage_cog into their own respective files
