+++ wall_e/cogs/here.py
@@ -17,6 +17,7 @@ class Here(commands.Cog):
         self.logger = log_info[0]
         self.debug_log_file_absolute_path = log_info[1]
         self.error_log_file_absolute_path = log_info[2]
+        self.logger.info("[Here __init__()] initializing Here")
         self.bot = bot
         self.config = config
         self.guild = None
