+++ wall_e/cogs/mod.py
@@ -15,6 +15,7 @@ class Mod(commands.Cog):
         self.logger = log_info[0]
         self.debug_log_file_absolute_path = log_info[1]
         self.error_log_file_absolute_path = log_info[2]
+        self.logger.info("[Mod __init__()] initializing Mod")
         self.bot = bot
         self.config = config
         self.guild = None
