+++ wall_e/cogs/reminders.py
@@ -20,6 +20,7 @@ class Reminders(commands.Cog):
         self.logger = log_info[0]
         self.debug_log_file_absolute_path = log_info[1]
         self.error_log_file_absolute_path = log_info[2]
+        self.logger.info("[Reminders __init__()] initializing Reminders")
         self.bot = bot
         self.config = config
         self.guild = None
