+++ wall_e/cogs/role_commands.py
@@ -24,6 +24,7 @@ class RoleCommands(commands.Cog):
         self.logger = log_info[0]
         self.debug_log_file_absolute_path = log_info[1]
         self.error_log_file_absolute_path = log_info[2]
+        self.logger.info("[RoleCommands __init__()] initializing RoleCommands")
         self.bot = bot
         self.config = config
         self.guild = None
