+++ wall_e/utilities/bot_channel_manager.py
@@ -21,8 +21,8 @@ log_positioning = {
     "here_error": 15,
     "leveling_debug": 16,
     "leveling_error": 17,
-    "manage_cog_debug": 18,
-    "manage_cog_error": 19,
+    "manage_test_guild_debug": 18,
+    "manage_test_guild_error": 19,
     "misc_debug": 20,
     "misc_error": 21,
     "mod_debug": 22,
