+++ wall_e/utilities/config/local.ini
@@ -43,7 +43,7 @@ frosh = 0
 health_checks = 1
 here = 1
 leveling = 1
-manage_cog = 1
+manage_test_guild = 0
 misc = 1
 mod = 1
 reminders = 1
